move secure_store
